Netflix Movies
Netflix, one of the largest streaming platforms in the world, offers a vast library of movies across various genres, languages, and countries. This project aims to analyze the Netflix movie and Tv Shows dataset to uncover insights about the content available, trends in movie production and ratings, and user preferences.